requirements.txt
paho-mqtt==1.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
